More lint

+++ app/src/main/java/com/gputreats/orbitalexplorer/OrbitalTextures.java
@@ -56,13 +56,13 @@ class OrbitalTextures {
             // Calculate radius info
             float quadratureRadius = (float) radialFunction.getMaximumRadius();
             float maxLateral = quadratureData[quadratureData.length - 2];
-            float maximumRadius = (float) Math.sqrt(quadratureRadius * quadratureRadius
-                    + maxLateral * maxLateral);
+            float maximumRadius = (float) Math.sqrt((double) (quadratureRadius * quadratureRadius
+                    + maxLateral * maxLateral));
 
             // Load new radial texture
             final int radialTextureSize = 1024;
             float[] radialData = functionToBuffer2(radialFunction.getOscillatingPart(),
-                    0.0, maximumRadius, radialTextureSize);
+                    0.0, (double) maximumRadius, radialTextureSize);
             radialTexture.bindToTexture2DAndSetImage(radialTextureSize, 1, radialData);
 
             // Load new azimuthal texture
